user_service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
pymongo==4.6.1
stomp.py==8.1.2